Move form related classes to forms.py
