+++ carrent/carrentapp/models.py
@@ -1,7 +1,7 @@
 from django.db import models
-from django.contrib.auth.models import AbstractBaseUser, BaseUserManager, PermissionsMixin
-
+from django.contrib.auth.models import AbstractBaseUser, BaseUserManager
 # Create your models here.
+
 
 class AccountManagerCustom(BaseUserManager):
